+++ src/bronze/sql/create_table_laps.sql
@@ -4,15 +4,15 @@ CREATE TABLE laps (
     duration_sector_1 DOUBLE,
     duration_sector_2 DOUBLE,
     duration_sector_3 DOUBLE,
-    i1_speed BIGINT,
-    i2_speed BIGINT,
+    i1_speed DOUBLE,
+    i2_speed DOUBLE,
     is_pit_out_lap BOOLEAN,
     lap_duration DOUBLE,
     lap_number BIGINT,
-    meeting_key BIGINT,
+    meeting_key INT,
     segments_sector_1 ARRAY<BIGINT>,
     segments_sector_2 ARRAY<BIGINT>,
     segments_sector_3 ARRAY<BIGINT>,
-    session_key BIGINT,
-    st_speed BIGINT
+    session_key INT,
+    st_speed DOUBLE
 )
