+++ src/bronze/sql/create_table_pits.sql
@@ -3,8 +3,8 @@ CREATE TABLE pits (
     driver_number BIGINT,
     lane_duration DOUBLE,
     lap_number BIGINT,
-    meeting_key BIGINT,
-    session_key BIGINT,
+    meeting_key INT,
+    session_key INT,
     pit_duration DOUBLE,
     stop_duration DOUBLE
 )
