+++ src/bronze/sql/create_table_position.sql
@@ -1,7 +1,7 @@
 CREATE TABLE position (
     date STRING,
     driver_number BIGINT,
-    meeting_key BIGINT,
+    meeting_key INT,
     position BIGINT,
-    session_key BIGINT
+    session_key INT
 )
